backend-python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Document processing
pypdf2==3.0.1
openpyxl==3.1.2
python-docx==1.0.1
pandas==2.1.3

# ML/AI dependencies
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
